Backing away from a full refactor to get this working.
